app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
oci==2.150.3
python-docx==1.1.2
pydantic==2.10.4
